Day Zero Experience: Update Helm Chart to remove UAA endpoint & use StorageClass

+++ deploy/db/scripts/run-postflight-job.k8s.sh
@@ -65,6 +65,10 @@ PGSQL_PASSWORD=$PWD goose --env=k8s dbversion
 echo "Database operation(s) complete."
 
 
+# Check if Upgrade Lock file exists
+if [ ! -f "/$UPGRADE_VOLUME/$UPGRADE_LOCK_FILENAME" ]; then
+  exit 1
+fi
 # Remove the lock file on the shared volume
 echo "Removing the $UPGRADE_LOCK_FILENAME file from the shared upgrade volume $UPGRADE_VOLUME."
 rm /$UPGRADE_VOLUME/$UPGRADE_LOCK_FILENAME || true
